- menambahkan manual dependency injection untuk repository dan dao - peningkatan fitur autentikasi dengan menambahkan pesan error pada edittext di login dan register activity apabila input pengguna tidak memenuhi kriteria - mengimplementasi fitur untuk menyimpan data login pengguna menggunakan userpreference / datastore.

+++ app/src/main/java/com/telyu/nourimate/data/local/models/User.kt
@@ -8,7 +8,7 @@ import kotlinx.parcelize.Parcelize
 
 @Entity (tableName = "users")
 @Parcelize
-data class User (
+data class User(
     @PrimaryKey(autoGenerate = false)
     @field:ColumnInfo(name = "name")
     val name: String,
@@ -17,7 +17,7 @@ data class User (
     val email: String,
 
     @field:ColumnInfo(name = "phone_number")
-    val phoneNumber: Int = 0,
+    val phoneNumber: Long = 0,
 
     @field:ColumnInfo(name = "password")
     val password: String,
